+++ PolarAlignment/Vector3.cs
@@ -52,7 +52,7 @@ namespace NINA.Plugins.PolarAlignment {
                 double temperature = refrectionParameters.Temperature;
                 double relativeHumidity = refrectionParameters.RelativeHumidity;
                 double elevation = refrectionParameters.Elevation;
-                const double wavelength = 0.55d;
+                double wavelength = refrectionParameters.Wavelength;
                 Logger.Info($"Transforming coordinates with refraction parameters. Pressure={pressurehPa}, Temperature={temperature}, Humidity={relativeHumidity}, Wavelength={wavelength}");
                 topo = coordinates.Transform(latitude, longitude, elevation, pressurehPa, temperature, relativeHumidity, wavelength);
             } else {