+++ PolarAlignment/Changelog.md
@@ -3,6 +3,7 @@
 ## Version 1.7.1.0
 - Add an option to continue tracking when TPPA is done. Use with caution to not run into pier collisions!
 - Prepopulate the filter with the platesolving filter for defaults
+- When refraction correction is enabled, the pole will now also be corrected for it to determine the initial error
 
 ## Version 1.7.0.0
 - Show a loading spinner while a new image is waiting for a solve to update the error details. The spinner is shown in the total error details.